Bugs Fixing - fixed double counting when calculating probabilities - fixed bug in PopsUnit that produced bad reports - fixed bug in CohortsUnit that made the program generate reports for cohorts with 0 personnel
Reworked CIPsUnit and ReportUnit.

Renamed OutputData to DailyReport and added new function.

Added basic functionality for DataService.

+++ CBRN_Project/CBRN_Project/MVVM/Models/Chemical Model/ChemExIcon.cs
@@ -6,6 +6,6 @@ namespace CBRN_Project.MVVM.Models.Chemical
     {
         public Icon                         Icon    { get; set; }
         public Dictionary<string, double>   Pbties  { get; set; }
-        public Dictionary<string, double>   Pops    { get; set; }
+        public Dictionary<string, int>      Pops    { get; set; }
     }
 }
